+++ beast-of-argh/beast-of-argh.d/local/share/cook/bin/configure-loki.sh
@@ -20,23 +20,32 @@ mkdir -p /tmp/loki
 
 # loki setup
 
-# create loki user
-/usr/sbin/pw useradd -n loki -c 'loki user' -m -s /usr/sbin/nologin -h -
+# removed as pkg install should cover it (2024-07-08)
+## create loki user
+#/usr/sbin/pw useradd -n loki -c 'loki user' -m -s /usr/sbin/nologin -h -
 
 # shellcheck disable=SC3003,SC2039
 # safe(r) separator for sed
 sep=$'\001'
 
-# copy in loki rc
-cp -f "$TEMPLATEPATH/loki.rc.in" /usr/local/etc/rc.d/loki
-chmod +x /usr/local/etc/rc.d/loki
+# removed as pkg install should cover it (2024-07-08)
+## copy in loki rc
+#cp -f "$TEMPLATEPATH/loki.rc.in" /usr/local/etc/rc.d/loki
+#chmod +x /usr/local/etc/rc.d/loki
+
 service loki enable || true
-sysrc loki_syslog_output_enable="YES" || true
+
+# this is invalid  on pkg install grafana-loki
+# logs are in /var/log/loki/loki.log
+#sysrc loki_syslog_output_enable="YES" || true
+
 # not specifically needed if permissions on /tmp/loki and loki-local-config.yaml
 # are correct
 #sysrc loki_user="loki"
 #sysrc loki_group="loki"
-#sysrc loki_config="/usr/local/etc/loki-local-config.yaml"
+
+# add this in for pkg install of grafana-loki to allow legacy setup stay same (2024-07-08)
+sysrc loki_config="/usr/local/etc/loki-local-config.yaml"
 
 # copy in loki config file, make sure loki can RW it
 < "$TEMPLATEPATH/loki-local-config.yaml.in" \
@@ -50,11 +59,31 @@ chown -R loki:loki /tmp/loki
 
 # promtail setup
 
+# promtail needs a directory it has permission to write to
+mkdir -p /mnt/promtail
+
+# promtail default group is loki
+chown -R promtail:loki /mnt/promtail
+
+# removed as pkg install should cover it (2024-07-08)
 # copy in the promtail rc file
-cp "$TEMPLATEPATH/promtail.rc.in" /usr/local/etc/rc.d/promtail
-chmod +x /usr/local/etc/rc.d/promtail
+#cp "$TEMPLATEPATH/promtail.rc.in" /usr/local/etc/rc.d/promtail
+#chmod +x /usr/local/etc/rc.d/promtail
+
 service promtail enable || true
-sysrc promtail_syslog_output_enable="YES" || true
+
+# promtail needs to run with root perms because syslog-ng is saving
+# log files to /mnt/logs/remote and the files have 600 perms
+# disabled #
+# trying with read perms configured in syslog-ng instead
+# sysrc promtail_user="root"
+
+# invalid with pkg install grafana-loki
+# logs are in /var/log/promtail/promtail.log
+#sysrc promtail_syslog_output_enable="YES" || true
+
+# add this in for pkg install of grafana-loki to allow legacy setup stay same (2024-07-08)
+sysrc promtail_config="/usr/local/etc/promtail-local-config.yaml"
 
 # copy in the promtail config file
 < "$TEMPLATEPATH/promtail-local-config.yaml.in" \