+++ mailhub-potluck/mailhub-potluck.sh
@@ -296,7 +296,8 @@ step "Pull files"
 #git pull --depth=1 origin 2023Q2
 #git pull --depth=1 origin 2023Q3
 #git pull --depth=1 origin 2024Q1
-git pull --depth=1 origin 2024Q2
+#git pull --depth=1 origin 2024Q2
+git pull --depth=1 origin 2024Q3
 
 #step "Port build openssl, remove existing, replace with this port"
 ##required for latest / main branch