+++ nomad-server/CHANGELOG.md
@@ -1,3 +1,9 @@
+3.17
+
+* Version bump for new base image
+
+---
+
 3.16
 
 * Version bump for new base image