+++ wordpress-nginx-nomad/CHANGELOG.md
@@ -1,3 +1,10 @@
+2.14
+
+* Version bump for new base image
+* Python is now py311
+
+---
+
 2.13
 
 * Version bump for new base image